backend/requirements.txt
django

django-cors-headers


djangorestframework


PyJWT 
mysql-connector-python
mysqlclient 
djangorestframework-simplejwt
dj_database_url

PyMySQL==1.0.2 

psycopg2-binary

Pillow